# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeshRadioMapResultSet": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "kind": {
            "const": "mesh_radio_map",
            "default": "mesh_radio_map",
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "metric": {
            "default": "rss_dbm",
            "enum": [
              "path_gain_db",
              "rss_dbm"
            ],
            "title": "Metric",
            "type": "string"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "simulation_config_id": {
            "title": "Simulation Config Id",
            "type": "string"
          },
          "surfaces": {
            "items": {
              "$ref": "#/components/schemas/MeshRadioMapSurface"
            },
            "title": "Surfaces",
            "type": "array"
          },
          "tx_id": {
            "title": "Tx Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "backend",
          "simulation_config_id",
          "tx_id"
        ],
        "title": "MeshRadioMapResultSet",
        "type": "object"
      },
...
      "PathResultSet": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "kind": {
            "const": "paths",
            "default": "paths",
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "paths": {
            "items": {
              "$ref": "#/components/schemas/RayPath"
            },
            "title": "Paths",
            "type": "array"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "simulation_config_id": {
            "title": "Simulation Config Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "backend",
          "simulation_config_id"
        ],
        "title": "PathResultSet",
        "type": "object"
      },
...
      "RadioMapResultSet": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "grid": {
            "$ref": "#/components/schemas/RadioMapGrid"
          },
          "kind": {
            "const": "radio_map",
            "default": "radio_map",
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "metric": {
            "default": "rss_dbm",
            "enum": [
              "path_gain_db",
              "rss_dbm",
              "sinr_db"
            ],
            "title": "Metric",
            "type": "string"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "serving_tx": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Tx"
          },
          "simulation_config_id": {
            "title": "Simulation Config Id",
            "type": "string"
          },
          "tx_id": {
            "title": "Tx Id",
            "type": "string"
          },
          "tx_ids": {
            "items": {
              "type": "string"
            },
            "title": "Tx Ids",
            "type": "array"
          },
          "values": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Values",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "backend",
          "simulation_config_id",
          "tx_id",
          "grid",
          "values"
        ],
        "title": "RadioMapResultSet",
        "type": "object"
      },
...
      "TrajectoryResultSet": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "kind": {
            "const": "trajectory",
            "default": "trajectory",
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TrajectorySample"
            },
            "title": "Samples",
            "type": "array"
          },
          "simulation_config_id": {
            "title": "Simulation Config Id",
            "type": "string"
          },
          "ue_id": {
            "title": "Ue Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "backend",
          "simulation_config_id",
          "ue_id"
        ],
        "title": "TrajectoryResultSet",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Unified RF-visual scene authoring, RF material assignment, Sionna RT projection compilation, and simulation result APIs.",
    "title": "SionnaTwin Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/results/mesh-radio-map": {
      "get": {
        "operationId": "get_mesh_radio_map_result_api_projects__project_id__results_mesh_radio_map_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "result_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeshRadioMapResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mesh Radio Map Result",
        "tags": [
          "simulate"
        ]
      }
    },
    "/api/projects/{project_id}/results/paths": {
      "get": {
        "operationId": "get_paths_result_api_projects__project_id__results_paths_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "result_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Paths Result",
        "tags": [
          "simulate"
        ]
      }
    },
    "/api/projects/{project_id}/results/radio-map": {
      "get": {
        "operationId": "get_radio_map_result_api_projects__project_id__results_radio_map_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "result_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadioMapResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Radio Map Result",
        "tags": [
          "simulate"
        ]
      }
    },
...
    "/api/projects/{project_id}/results/trajectory": {
      "get": {
        "operationId": "get_trajectory_result_api_projects__project_id__results_trajectory_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "result_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trajectory Result",
        "tags": [
          "simulate"
        ]
      }
    },
...
    "/api/projects/{project_id}/simulate/beamforming": {
      "post": {
        "description": "MIMO beamforming gain (MRT / SVD) over one TX->RX link. Computed on\ndemand and returned directly (not stored as a result set).",
        "operationId": "simulate_beamforming_api_projects__project_id__simulate_beamforming_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BeamformingRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeamformingResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Beamforming",
        "tags": [
          "simulate"
        ]
      }
    },
    "/api/projects/{project_id}/simulate/mesh-radio-map": {
      "post": {
        "description": "Per-triangle coverage on the requested prims' surfaces (facades,\nroads, floors) - probe receivers at triangle centers, chunk-solved with\nthe active backend.",
        "operationId": "simulate_mesh_radio_map_api_projects__project_id__simulate_mesh_radio_map_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeshRadioMapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeshRadioMapResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Mesh Radio Map",
        "tags": [
          "simulate"
        ]
      }
    },
    "/api/projects/{project_id}/simulate/paths": {
      "post": {
        "operationId": "simulate_paths_api_projects__project_id__simulate_paths_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SimulateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Paths",
        "tags": [
          "simulate"
        ]
      }
    },
    "/api/projects/{project_id}/simulate/radio-map": {
      "post": {
        "operationId": "simulate_radio_map_api_projects__project_id__simulate_radio_map_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SimulateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadioMapResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Radio Map",
        "tags": [
          "simulate"
        ]
      }
    },
...
    "/api/projects/{project_id}/simulate/trajectory": {
      "post": {
        "operationId": "simulate_trajectory_api_projects__project_id__simulate_trajectory_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrajectorySimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryResultSet"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Trajectory",
        "tags": [
          "simulate"
        ]
      }
    }
  }
}